Add gan_loss_functions and seperate networks from model
