+++ src/configs/template_GAN_train_config.json
@@ -1,6 +1,6 @@
 {
     "n_gpu": 1,
-    "optimizer_G": {
+    "optimizer_generator": {
         "type": "Adam",
         "args": {
             "lr": 0.0001,
@@ -8,7 +8,7 @@
             "amsgrad": true
         }
     },
-    "optimizer_D": {
+    "optimizer_discriminator": {
         "type": "Adam",
         "args": {
             "lr": 0.0001,
@@ -49,7 +49,7 @@
         "type": "MnistDataLoader",
         "args": {
             "data_dir": "../datasets/mnist",
-            "batch_size": 32,
+            "batch_size": 16,
             "validation_split": 0.1,
             "num_workers": 4,
             "img_size": 64,
@@ -57,14 +57,16 @@
         }
     },
     "losses": {
-        "0": {
+    },
+    "gan_losses": {
+        "generator": {
             "type": "HingeLossG",
             "args": {
                 "nickname": "hinge_loss_generator",
                 "weight": 1
             }
         },
-        "1": {
+        "discriminator": {
             "type": "HingeLossD",
             "args": {
                 "nickname": "hinge_loss_discriminator",
