+++ src/model/model.py
@@ -5,9 +5,9 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(__file__))))  # 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from torch.nn.utils import spectral_norm
 
-from model.base_model import BaseModel
+from .base_model import BaseModel
+from .networks import MnistGenerator, MnistDiscriminator
 
 
 class MnistModel(BaseModel):
@@ -36,71 +36,6 @@ class MnistModel(BaseModel):
         }
 
 
-def normal_init(m, mean, std):
-    if isinstance(m, nn.ConvTranspose2d) or isinstance(m, nn.Conv2d):
-        m.weight.data.normal_(mean, std)
-        m.bias.data.zero_()
-
-
-class MnistGenerator(nn.Module):
-    # architecture reference: https://github.com/znxlwm/pytorch-MNIST-CelebA-GAN-DCGAN/blob/master/pytorch_MNIST_DCGAN.py  # NOQA
-    def __init__(self, d=128):
-        super().__init__()
-        self.deconv1 = nn.ConvTranspose2d(100, d * 8, 4, 1, 0)
-        self.deconv1_bn = nn.BatchNorm2d(d * 8)
-        self.deconv2 = nn.ConvTranspose2d(d * 8, d * 4, 4, 2, 1)
-        self.deconv2_bn = nn.BatchNorm2d(d * 4)
-        self.deconv3 = nn.ConvTranspose2d(d * 4, d * 2, 4, 2, 1)
-        self.deconv3_bn = nn.BatchNorm2d(d * 2)
-        self.deconv4 = nn.ConvTranspose2d(d * 2, d, 4, 2, 1)
-        self.deconv4_bn = nn.BatchNorm2d(d)
-        self.deconv5 = nn.ConvTranspose2d(d, 1, 4, 2, 1)
-
-    def weight_init(self, mean, std):
-        for m in self._modules:
-            normal_init(self._modules[m], mean, std)
-
-    def forward(self, input):
-        # x = F.relu(self.deconv1(input))
-        x = F.relu(self.deconv1_bn(self.deconv1(input)))
-        x = F.relu(self.deconv2_bn(self.deconv2(x)))
-        x = F.relu(self.deconv3_bn(self.deconv3(x)))
-        x = F.relu(self.deconv4_bn(self.deconv4(x)))
-        x = F.tanh(self.deconv5(x))
-        return x
-
-
-class MnistDiscriminator(nn.Module):
-    # architecture reference: https://github.com/znxlwm/pytorch-MNIST-CelebA-GAN-DCGAN/blob/master/pytorch_MNIST_DCGAN.py  # NOQA
-    def __init__(self, d=128, spectral_normalization=True):
-        super().__init__()
-        self.conv1 = nn.Conv2d(1, d, 4, 2, 1)
-        self.conv2 = nn.Conv2d(d, d * 2, 4, 2, 1)
-        self.conv2_bn = nn.BatchNorm2d(d * 2)
-        self.conv3 = nn.Conv2d(d * 2, d * 4, 4, 2, 1)
-        self.conv3_bn = nn.BatchNorm2d(d * 4)
-        self.conv4 = nn.Conv2d(d * 4, d * 8, 4, 2, 1)
-        self.conv4_bn = nn.BatchNorm2d(d * 8)
-        self.conv5 = nn.Conv2d(d * 8, 1, 4, 1, 0)
-
-        if spectral_normalization:
-            for attr_name in [f'conv{i}' for i in range(1, 6)]:
-                new_attr = spectral_norm(getattr(self, attr_name))
-                setattr(self, attr_name, new_attr)
-
-    def weight_init(self, mean, std):
-        for m in self._modules:
-            normal_init(self._modules[m], mean, std)
-
-    def forward(self, input):
-        x = F.leaky_relu(self.conv1(input), 0.2)
-        x = F.leaky_relu(self.conv2_bn(self.conv2(x)), 0.2)
-        x = F.leaky_relu(self.conv3_bn(self.conv3(x)), 0.2)
-        x = F.leaky_relu(self.conv4_bn(self.conv4(x)), 0.2)
-        x = F.sigmoid(self.conv5(x))
-        return x
-
-
 class MnistGAN(BaseModel):
     def __init__(self, spectral_normalization=True):
         super().__init__()
@@ -110,20 +45,19 @@ class MnistGAN(BaseModel):
         self.generator.weight_init(mean=0.0, std=0.02)
         self.discriminator.weight_init(mean=0.0, std=0.02)
 
-    def forward(self, data_dict, model='G'):
+    def forward(self, data_dict, network_name):
         x = data_dict['data_input']
         batch_size = x.size(0)
         z = torch.randn((batch_size, 100)).view(-1, 100, 1, 1).to(x.device)
         G_z = self.generator(z)
+        D_G_z = self.discriminator(G_z).squeeze()
         model_output = {
-            "G_z": G_z
+            "G_z": G_z,
+            "D_G_z": D_G_z
         }
-        if model == 'D':
-            D_G_z = self.discriminator(G_z).squeeze()
+        if network_name == 'discriminator':
             D_x = self.discriminator(x).squeeze()
-
-            model["D_G_z"] = D_G_z
-            model["D_x"] = D_x
+            model_output["D_x"] = D_x
         return model_output
 
     @property
