+++ src/model/networks.py
@@ -1,0 +1,68 @@
+import torch.nn as nn
+import torch.nn.functional as F
+from torch.nn.utils import spectral_norm
+
+
+def normal_init(m, mean, std):
+    if isinstance(m, nn.ConvTranspose2d) or isinstance(m, nn.Conv2d):
+        m.weight.data.normal_(mean, std)
+        m.bias.data.zero_()
+
+
+class MnistGenerator(nn.Module):
+    # architecture reference: https://github.com/znxlwm/pytorch-MNIST-CelebA-GAN-DCGAN/blob/master/pytorch_MNIST_DCGAN.py  # NOQA
+    def __init__(self, d=128):
+        super().__init__()
+        self.deconv1 = nn.ConvTranspose2d(100, d * 8, 4, 1, 0)
+        self.deconv1_bn = nn.BatchNorm2d(d * 8)
+        self.deconv2 = nn.ConvTranspose2d(d * 8, d * 4, 4, 2, 1)
+        self.deconv2_bn = nn.BatchNorm2d(d * 4)
+        self.deconv3 = nn.ConvTranspose2d(d * 4, d * 2, 4, 2, 1)
+        self.deconv3_bn = nn.BatchNorm2d(d * 2)
+        self.deconv4 = nn.ConvTranspose2d(d * 2, d, 4, 2, 1)
+        self.deconv4_bn = nn.BatchNorm2d(d)
+        self.deconv5 = nn.ConvTranspose2d(d, 1, 4, 2, 1)
+
+    def weight_init(self, mean, std):
+        for m in self._modules:
+            normal_init(self._modules[m], mean, std)
+
+    def forward(self, input):
+        # x = F.relu(self.deconv1(input))
+        x = F.relu(self.deconv1_bn(self.deconv1(input)))
+        x = F.relu(self.deconv2_bn(self.deconv2(x)))
+        x = F.relu(self.deconv3_bn(self.deconv3(x)))
+        x = F.relu(self.deconv4_bn(self.deconv4(x)))
+        x = F.tanh(self.deconv5(x))
+        return x
+
+
+class MnistDiscriminator(nn.Module):
+    # architecture reference: https://github.com/znxlwm/pytorch-MNIST-CelebA-GAN-DCGAN/blob/master/pytorch_MNIST_DCGAN.py  # NOQA
+    def __init__(self, d=128, spectral_normalization=True):
+        super().__init__()
+        self.conv1 = nn.Conv2d(1, d, 4, 2, 1)
+        self.conv2 = nn.Conv2d(d, d * 2, 4, 2, 1)
+        self.conv2_bn = nn.BatchNorm2d(d * 2)
+        self.conv3 = nn.Conv2d(d * 2, d * 4, 4, 2, 1)
+        self.conv3_bn = nn.BatchNorm2d(d * 4)
+        self.conv4 = nn.Conv2d(d * 4, d * 8, 4, 2, 1)
+        self.conv4_bn = nn.BatchNorm2d(d * 8)
+        self.conv5 = nn.Conv2d(d * 8, 1, 4, 1, 0)
+
+        if spectral_normalization:
+            for attr_name in [f'conv{i}' for i in range(1, 6)]:
+                new_attr = spectral_norm(getattr(self, attr_name))
+                setattr(self, attr_name, new_attr)
+
+    def weight_init(self, mean, std):
+        for m in self._modules:
+            normal_init(self._modules[m], mean, std)
+
+    def forward(self, input):
+        x = F.leaky_relu(self.conv1(input), 0.2)
+        x = F.leaky_relu(self.conv2_bn(self.conv2(x)), 0.2)
+        x = F.leaky_relu(self.conv3_bn(self.conv3(x)), 0.2)
+        x = F.leaky_relu(self.conv4_bn(self.conv4(x)), 0.2)
+        x = F.sigmoid(self.conv5(x))
+        return x
